* Clear the DelegatorChunk's when clearing Delegator, so it doesn't throw on some obscure bug. Not going to be looking for this one anyway as there is a rewrite scheduled.
* Added "tos = default|lowdelay|throughput|reliability|mincost" option.

* Made EntryList::value_type a pointer, fixed so FileMeta is
initialized and path set before Download::open().


git-svn-id: svn://rakshasa.no/libtorrent/trunk/rtorrent@662 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/display/window_download_chunks_seen.cc
@@ -80,7 +80,7 @@ WindowDownloadChunksSeen::redraw() {
   const uint8_t* last = seen + m_download->download()->chunks_total();
 
   for (int y = 1; y < m_canvas->get_height() && chunk < last; ++y) {
-    position = buffer + std::max(snprintf(buffer, end - buffer, "%5d", chunk - seen), 0);
+    position = buffer + std::max(snprintf(buffer, end - buffer, "%5d", (int)(chunk - seen)), 0);
 
     while (chunk < last) {
       if ((chunk - seen) % 10 == 0) {
